Refactor GUI packages
